Added base view model module.
